Conecta useSeeds com as seeds do front end
