+++ seeds/students.js
@@ -1,193 +1,8 @@
-const Student = require('../server/models/Student')
+const { students } = require('../test/seeds')
 
 exports.seed = async function(knex, Promise) {
   // Deletes ALL existing entries
   await knex('students').del()
 
-  const data = [
-    {
-      name: 'FELIPE SOUZA FERREIRA',
-      registrationNumber: '201704940001',
-      crg: 8.3,
-      course: 'cbcc',
-      email: 'notgoo@gmiau.com',
-      isFit: false,
-      isConcluding: false,
-      isActive: true,
-      isForming: true,
-      isGraduating: false,
-      academicHighlight: false,
-      cancelled: false,
-      prescribed: false,
-      mailingList: 'none',
-      mailingListToRemove: 'none',
-      mailingListToAdd: 'active',
-      entryDate: '05-28-2019',
-      advisor: 'Orientador Válido',
-      defenseDate: null,
-      term: null,
-      recordSigned: false,
-      termPaper: false,
-      cd: false,
-      isUndergraduate: false
-    },
-    {
-      name: 'LAURA CARDOSO CASTRO',
-      registrationNumber: '201304940002',
-      crg: 6.63,
-      course: 'cbsi',
-      email: 'youngboy@yahoo.com',
-      isFit: false,
-      isConcluding: false,
-      isActive: true,
-      isForming: false,
-      isGraduating: true,
-      academicHighlight: false,
-      cancelled: false,
-      prescribed: false,
-      mailingList: 'none',
-      mailingListToRemove: 'none',
-      mailingListToAdd: 'active',
-      entryDate: '05-27-2019',
-      advisor: null,
-      defenseDate: null,
-      term: null,
-      recordSigned: false,
-      termPaper: false,
-      cd: false,
-      isUndergraduate: false
-    },
-    {
-      name: 'JOSE FERREIRA SILVA',
-      registrationNumber: '200504940003',
-      crg: 8.3,
-      course: 'cbsi',
-      email: 'naosouativo@gmail.com',
-      isFit: false,
-      isConcluding: true,
-      isActive: false,
-      isForming: false,
-      isGraduating: false,
-      academicHighlight: false,
-      cancelled: false,
-      prescribed: false,
-      mailingList: 'active',
-      mailingListToRemove: 'active',
-      mailingListToAdd: 'concluding',
-      entryDate: '01-28-2019',
-      advisor: 'Mais um orientador',
-      defenseDate: null,
-      term: null,
-      recordSigned: false,
-      termPaper: false,
-      cd: false,
-      isUndergraduate: false
-    },
-    {
-      name: 'ENZO FERREIRA ALVES',
-      registrationNumber: '201104940004',
-      crg: 7.89,
-      course: 'cbcc',
-      email: 'slug@gmail.com',
-      isFit: true,
-      isConcluding: false,
-      isActive: true,
-      isForming: false,
-      isGraduating: true,
-      academicHighlight: false,
-      cancelled: false,
-      prescribed: false,
-      mailingList: 'none',
-      mailingListToRemove: 'none',
-      mailingListToAdd: 'active',
-      entryDate: '04-20-2019',
-      advisor: null,
-      defenseDate: null,
-      term: null,
-      recordSigned: false,
-      termPaper: false,
-      cd: false,
-      isUndergraduate: false
-    },
-    {
-      name: 'KAUAN CARVALHO SANTOS',
-      registrationNumber: '200604940005',
-      crg: 7.897,
-      course: 'cbcc',
-      email: 'dofuturo@gmail.com.br',
-      isFit: false,
-      isConcluding: false,
-      isActive: false,
-      isForming: false,
-      isGraduating: false,
-      academicHighlight: false,
-      cancelled: true,
-      prescribed: false,
-      mailingList: 'active',
-      mailingListToRemove: 'active',
-      mailingListToAdd: 'none',
-      entryDate: '05-18-2018',
-      advisor: null,
-      defenseDate: null,
-      term: null,
-      recordSigned: false,
-      termPaper: false,
-      cd: false,
-      isUndergraduate: false
-    },
-    {
-      name: 'EDUARDO ALVES LIMA',
-      registrationNumber: '201804940006',
-      crg: 6.001,
-      course: 'cbcc',
-      email: 'oldfashioned33@hotmail.com',
-      isFit: false,
-      isConcluding: false,
-      isActive: true,
-      isForming: false,
-      isGraduating: false,
-      academicHighlight: false,
-      cancelled: false,
-      prescribed: false,
-      mailingList: 'none',
-      mailingListToRemove: 'none',
-      mailingListToAdd: 'active',
-      entryDate: null,
-      advisor: null,
-      defenseDate: null,
-      term: null,
-      recordSigned: false,
-      termPaper: false,
-      cd: false,
-      isUndergraduate: false
-    },
-    {
-      name: 'JULIAN BARBOSA SANTOS',
-      registrationNumber: '199604940007',
-      crg: 8.3,
-      course: 'cbcc',
-      email: 'uneccessary@ufpa.br',
-      isFit: false,
-      isConcluding: true,
-      isActive: false,
-      isForming: false,
-      isGraduating: false,
-      academicHighlight: false,
-      cancelled: false,
-      prescribed: false,
-      mailingList: 'active',
-      mailingListToRemove: 'active',
-      mailingListToAdd: 'concluding',
-      entryDate: '02-01-2019',
-      advisor: null,
-      defenseDate: null,
-      term: null,
-      recordSigned: false,
-      termPaper: false,
-      cd: false,
-      isUndergraduate: false
-    }
-  ]
-
-  return Promise.all(data.map(student => Student.forge(student).save()))
+  return students()
 }
